backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
deep-translator
pydantic
langchain
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb